% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dForm.R
\name{dForm}
\alias{dForm}
\title{R6 Class for downloading, caching, and basic processing of SEC Form D data}
\description{
Downloads, caches, combines, and optionally de-duplicates SEC Form D data
}
\details{
A dForm object can download Form D data, make that data available in your session, and optionally de-duplicate that data
using previous accessions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{codebook}}{Form D data codebook}

\item{\code{submissions}}{Combined submission data for selected years and quarters}

\item{\code{issuers}}{Combined issuers data for selected years and quarters}

\item{\code{offerings}}{Combined offerings data for selected years and quarters}

\item{\code{recipients}}{Combined recipients data for selected years and quarters}

\item{\code{related_persons}}{Combined related persons data for selected years and quarters}

\item{\code{signatures}}{Combined signatures data for selected years and quarters}

\item{\code{previous_accessions}}{Combined previous accessions data for selected years and quarters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dForm-load_data}{\code{dForm$load_data()}}
\item \href{#method-dForm-aggregate_data}{\code{dForm$aggregate_data()}}
\item \href{#method-dForm-combine_data}{\code{dForm$combine_data()}}
\item \href{#method-dForm-get_codebook}{\code{dForm$get_codebook()}}
\item \href{#method-dForm-clone}{\code{dForm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dForm-load_data"></a>}}
\if{latex}{\out{\hypertarget{method-dForm-load_data}{}}}
\subsection{Method \code{load_data()}}{
Download and read Form D data for chosen years and quarters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dForm$load_data(
  years,
  quarter = c(1:4),
  remove_duplicates = TRUE,
  use_cache = TRUE,
  contact
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{years}}{The separator to use for value concatenation}

\item{\code{quarter}}{Quarters to download. Defaults to all quarters (1 to 4).}

\item{\code{remove_duplicates}}{If \code{TRUE}, the previous accession numbers in the \code{offerings} data will be used to de-duplicate the data}

\item{\code{use_cache}}{If \code{TRUE}, read data from cached downloads. Otherwise, download and load the data.}

\item{\code{contact}}{Requester identity (organization + contact email) passed to
the downloader and appended to the User-Agent. Required: SEC blocks
requests that do not declare a contact. No default.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self for method chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dForm-aggregate_data"></a>}}
\if{latex}{\out{\hypertarget{method-dForm-aggregate_data}{}}}
\subsection{Method \code{aggregate_data()}}{
Aggregate all Form D data sets that are not unique on acessionnumber.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dForm$aggregate_data(sep = ", ")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sep}}{The separator to use for value concatenation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self for method chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dForm-combine_data"></a>}}
\if{latex}{\out{\hypertarget{method-dForm-combine_data}{}}}
\subsection{Method \code{combine_data()}}{
Combine two or more component data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dForm$combine_data(
  tables_to_combine = c("submissions", "offerings", "issuers")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tables_to_combine}}{The name of the data set to return a codebook for, one of 'submissions', 'issuers',
'offerings', 'recipients', 'related_persons', 'signatures'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dForm-get_codebook"></a>}}
\if{latex}{\out{\hypertarget{method-dForm-get_codebook}{}}}
\subsection{Method \code{get_codebook()}}{
Return a data frame codebook for one data set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dForm$get_codebook(data_set_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_set_name}}{The name of the data set to return a codebook for, one of 'submissions', 'issuers',
'offerings', 'recipients', 'related_persons', 'signatures'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dForm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dForm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dForm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
